pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[project]
name = "bloom-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["pyyaml>=6.0", "matplotlib>=3.7", "numpy>=1.24"]

[project.scripts]
bloomagent = "src.cli:main"
